fix
git-svn-id: https://www.brockmann-consult.de/svn/os/beam-4.x/trunk@792 d90d2d48-2b33-0410-a973-e72f85796a02

+++ beam-gpf/src/main/java/org/esa/beam/framework/gpf/internal/RasterImpl.java
@@ -102,7 +102,7 @@ public class RasterImpl implements Tile {
     }
 
     public final void setSample(int x, int y, float v) {
-        setSample(x, y, v);
+        setSample(x, y, (double)v);
     }
 
     public final double getSampleDouble(int x, int y) {
